shap-e-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
imageio==2.31.1
tqdm==4.66.1
trimesh==3.23.5
git+https://github.com/openai/shap-e.git
pyyaml==6.0.2
